+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/pdf/helpers/DataManipulation.java
@@ -11,8 +11,6 @@ import uk.gov.hmcts.reform.pip.publication.services.models.templatemodels.sscsda
 import uk.gov.hmcts.reform.pip.publication.services.models.templatemodels.sscsdailylist.Hearing;
 import uk.gov.hmcts.reform.pip.publication.services.models.templatemodels.sscsdailylist.Sitting;
 
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -23,9 +21,6 @@ import java.util.concurrent.atomic.AtomicReference;
 
 @SuppressWarnings({"PMD.TooManyMethods"})
 public final class DataManipulation {
-    private static final String POSTCODE = "postCode";
-    private static final String COURT_HOUSE = "courtHouse";
-    private static final int MINUTES_PER_HOUR = 60;
     private static final String CASE_SEQUENCE_INDICATOR = "caseSequenceIndicator";
     private static final ObjectMapper MAPPER = new ObjectMapper();
     public static final String APPLICANT = "applicant";
@@ -45,8 +40,8 @@ public final class DataManipulation {
     }
 
     public static void manipulateCopListData(JsonNode artefact) {
-        formatRegionName(artefact);
-        formatRegionalJoh(artefact);
+        LocationHelper.formatRegionName(artefact);
+        LocationHelper.formatRegionalJoh(artefact);
 
         artefact.get("courtLists").forEach(courtList -> {
             courtList.get("courtHouse").get("courtRoom").forEach(courtRoom -> {
@@ -54,7 +49,7 @@ public final class DataManipulation {
                     ((ObjectNode)session).put("formattedSessionJoh",
                                               DataManipulation.findAndManipulateJudiciaryForCop(session));
                     session.get("sittings").forEach(sitting -> {
-                        DataManipulation.calculateDuration(sitting);
+                        DateHelper.calculateDuration(sitting);
                         DataManipulation.findAndConcatenateHearingPlatform(sitting, session);
 
                         sitting.get("hearing").forEach(hearing -> {
@@ -66,63 +61,14 @@ public final class DataManipulation {
         });
     }
 
-    public static List<String> formatVenueAddress(JsonNode artefact) {
-        List<String> address = new ArrayList<>();
-        JsonNode arrayNode = artefact.get("venue").get("venueAddress").get("line");
-        for (JsonNode jsonNode : arrayNode) {
-            if (!jsonNode.asText().isEmpty()) {
-                address.add(jsonNode.asText());
-            }
-        }
-        if (!GeneralHelper.findAndReturnNodeText(artefact.get("venue").get("venueAddress"), POSTCODE).isEmpty()) {
-            address.add(artefact.get("venue").get("venueAddress").get(POSTCODE).asText());
-        }
-        return address;
-
-    }
-
-    public static void formatCourtAddress(JsonNode artefact) {
-        artefact.get("courtLists").forEach(courtList -> {
-            StringBuilder formattedCourtAddress = new StringBuilder();
-
-            if (courtList.get(COURT_HOUSE).has("courtHouseAddress")) {
-                JsonNode courtHouseAddress = courtList.get(COURT_HOUSE).get("courtHouseAddress");
-
-                GeneralHelper.loopAndFormatString(courtHouseAddress, "line",
-                                            formattedCourtAddress, "|");
-
-                checkAndFormatAddress(courtHouseAddress, "town",
-                                           formattedCourtAddress, '|');
-
-                checkAndFormatAddress(courtHouseAddress, "county",
-                                           formattedCourtAddress, '|');
-
-                checkAndFormatAddress(courtHouseAddress, POSTCODE,
-                                           formattedCourtAddress, '|');
-            }
-
-            ((ObjectNode)courtList.get(COURT_HOUSE)).put("formattedCourtHouseAddress",
-                formattedCourtAddress.toString().replaceAll(", $", ""));
-        });
-    }
-
-    private static void checkAndFormatAddress(JsonNode node, String nodeName,
-                                      StringBuilder builder, Character delimiter) {
-        if (!GeneralHelper.findAndReturnNodeText(node, nodeName).isEmpty()) {
-            builder
-                .append(node.get(nodeName).asText())
-                .append(delimiter);
-        }
-    }
-
     public static void manipulatedDailyListData(JsonNode artefact) {
         artefact.get("courtLists").forEach(courtList -> {
-            courtList.get(COURT_HOUSE).get("courtRoom").forEach(courtRoom -> {
+            courtList.get(LocationHelper.COURT_HOUSE).get("courtRoom").forEach(courtRoom -> {
                 courtRoom.get("session").forEach(session -> {
                     StringBuilder formattedJudiciary = new StringBuilder();
                     formattedJudiciary.append(findAndManipulateJudiciary(session));
                     session.get("sittings").forEach(sitting -> {
-                        calculateDuration(sitting);
+                        DateHelper.calculateDuration(sitting);
                         findAndConcatenateHearingPlatform(sitting, session);
 
                         sitting.get("hearing").forEach(hearing -> {
@@ -135,33 +81,10 @@ public final class DataManipulation {
                             hearing.get("case").forEach(DataManipulation::manipulateCaseInformation);
                         });
                     });
-                    formattedCourtRoomName(courtRoom, session, formattedJudiciary);
+                    LocationHelper.formattedCourtRoomName(courtRoom, session, formattedJudiciary);
                 });
             });
         });
-    }
-
-    public static void calculateDuration(JsonNode sitting) {
-        ZonedDateTime sittingStart = DateHelper.convertStringToUtc(sitting.get("sittingStart").asText());
-        ZonedDateTime sittingEnd = DateHelper.convertStringToUtc(sitting.get("sittingEnd").asText());
-
-        double durationAsHours = 0;
-        double durationAsMinutes = DateHelper.convertTimeToMinutes(sittingStart, sittingEnd);
-
-        if (durationAsMinutes >= MINUTES_PER_HOUR) {
-            durationAsHours = Math.floor(durationAsMinutes / MINUTES_PER_HOUR);
-            durationAsMinutes = durationAsMinutes - (durationAsHours * MINUTES_PER_HOUR);
-        }
-
-        String formattedDuration = DateHelper.formatDuration((int) durationAsHours,
-            (int) durationAsMinutes);
-
-        ((ObjectNode)sitting).put("formattedDuration", formattedDuration);
-
-        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("HH:mm");
-        String time = dtf.format(sittingStart);
-
-        ((ObjectNode)sitting).put("time", time);
     }
 
     private static void manipulateCaseInformation(JsonNode hearingCase) {
@@ -269,46 +192,6 @@ public final class DataManipulation {
         }
 
         return GeneralHelper.trimAnyCharacterFromStringEnd(formattedJudiciary.toString());
-    }
-
-    public static void formatRegionName(JsonNode artefact) {
-        try {
-            ((ObjectNode) artefact).put("regionName",
-                                        artefact.get("locationDetails").get("region").get("name").asText());
-        } catch (Exception e) {
-            ((ObjectNode) artefact).put("regionName", "");
-        }
-    }
-
-    public static void formatRegionalJoh(JsonNode artefact) {
-        StringBuilder formattedJoh = new StringBuilder();
-        try {
-            artefact.get("locationDetails").get("region").get("regionalJOH").forEach(joh -> {
-                if (formattedJoh.length() != 0) {
-                    formattedJoh.append(", ");
-                }
-
-                formattedJoh.append(GeneralHelper.findAndReturnNodeText(joh, "johKnownAs"));
-                formattedJoh.append(' ');
-                formattedJoh.append(GeneralHelper.findAndReturnNodeText(joh, "johNameSurname"));
-            });
-
-            ((ObjectNode) artefact).put("regionalJoh", formattedJoh.toString());
-        } catch (Exception e) {
-            ((ObjectNode) artefact).put("regionalJoh", "");
-        }
-    }
-
-    private static void formattedCourtRoomName(JsonNode courtRoom, JsonNode session,
-                                        StringBuilder formattedJudiciary) {
-        if (StringUtils.isBlank(formattedJudiciary.toString())) {
-            formattedJudiciary.append(courtRoom.get("courtRoomName").asText());
-        } else {
-            formattedJudiciary.insert(0, courtRoom.get("courtRoomName").asText() + ": ");
-        }
-
-        ((ObjectNode)session).put("formattedSessionCourtRoom",
-            GeneralHelper.trimAnyCharacterFromStringEnd(formattedJudiciary.toString()));
     }
 
     private static String findAndManipulateJudiciary(JsonNode session) {
